+++ ReleaseNotes.md
@@ -1,6 +1,12 @@
 # Release Notes
 
 ### 0.5.14
++ Launches animated buttons to special actions
+### Issues
++ Reset animated button style after be clicked 
++ Show the user name in some place in member area pages
+
+### 0.5.14
 + Spreads footer along all pages
 ### Issues
 + Show the user name in some place in member area pages